apps/requirements_main.txt
#conda create --name maindetect  python=3.10 -y
#sudo apt install ffmpeg
opencv-python
Pillow
requests